implement deepspeed library

+++ test_comm.py
@@ -13,7 +13,7 @@ my_device = torch.cuda.current_device()
 root_rank = 7
 
 # Load the seventh column from the file "log_0.txt"
-data = np.loadtxt('log_0.txt', usecols=6)
+data = np.loadtxt('log_0.txt', usecols=6, dtype=int)
 
 if my_rank == root_rank:
     print("my_rank " + str(my_rank) + "/" + str(world_size) + " my_device " + str(my_device) + "/" + str(torch.cuda.device_count()))
